Changed: Split up the CLI commands into separate modules.
